fixed db error and added function to get automatically latitude and longitude from an address

+++ www/create_tables.php
@@ -10,8 +10,8 @@
             stage_name VARCHAR(255) NOT NULL,
             stage_address VARCHAR(255) NOT NULL,
             stage_description TEXT,
-            stage_latitude DOUBLE NOT NULL,
-            stage_longitude DOUBLE NOT NULL,
+            stage_latitude DECIMAL(12, 10) NOT NULL,
+            stage_longitude DECIMAL(12, 10) NOT NULL,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
             updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
         );";
